Mellomlagre klage gir historikkinnslag
