+++ web/src/main/java/no/nav/foreldrepenger/web/app/tjenester/behandling/klage/aksjonspunkt/KlageFormKravLagreDto.java
@@ -1,0 +1,23 @@
+package no.nav.foreldrepenger.web.app.tjenester.behandling.klage.aksjonspunkt;
+
+import java.util.UUID;
+
+public interface KlageFormKravLagreDto {
+    boolean erKlagerPart();
+
+    boolean erFristOverholdt();
+
+    boolean erKonkret();
+
+    boolean erSignert();
+
+    UUID påKlagdBehandlingUuid();
+
+    boolean erTilbakekreving();
+
+    KlageTilbakekrevingDto klageTilbakekreving();
+
+    default UUID hentpåKlagdEksternBehandlingUuId() {
+        return erTilbakekreving() && klageTilbakekreving() != null ? klageTilbakekreving().tilbakekrevingUuid() : null;
+    }
+}
